+++ pkg/config/config.go
@@ -6,11 +6,13 @@ import (
 	"io/ioutil"
 	"os"
 
+	"github.com/go-kit/kit/log"
+	"github.com/go-kit/kit/log/level"
 	"github.com/weaveworks/common/server"
 
 	"github.com/drone/envsubst"
 	"github.com/grafana/agent/pkg/integrations"
-	"github.com/grafana/agent/pkg/loki"
+	"github.com/grafana/agent/pkg/logs"
 	"github.com/grafana/agent/pkg/prom"
 	"github.com/grafana/agent/pkg/tempo"
 	"github.com/grafana/agent/pkg/util"
@@ -30,9 +32,12 @@ var DefaultConfig = Config{
 type Config struct {
 	Server       server.Config              `yaml:"server,omitempty"`
 	Prometheus   prom.Config                `yaml:"prometheus,omitempty"`
-	Loki         loki.Config                `yaml:"loki,omitempty"`
 	Integrations integrations.ManagerConfig `yaml:"integrations,omitempty"`
 	Tempo        tempo.Config               `yaml:"tempo,omitempty"`
+
+	Logs               *logs.Config `yaml:"logs,omitempty"`
+	Loki               *logs.Config `yaml:"loki,omitempty"` // Deprecated: use Logs instead
+	UsedDeprecatedLoki bool         `yaml:"-"`
 
 	// We support a secondary server just for the /-/reload endpoint, since
 	// invoking /-/reload against the primary server can cause the server
@@ -52,10 +57,28 @@ func (c *Config) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	return unmarshal((*config)(c))
 }
 
+// LogDeprecations will log use of any deprecated fields to l as warn-level
+// messages.
+func (c *Config) LogDeprecations(l log.Logger) {
+	if c.UsedDeprecatedLoki {
+		level.Warn(l).Log("msg", "DEPRECATION NOTICE: `loki` is deprecated in favor of `logs`")
+	}
+}
+
 // ApplyDefaults sets default values in the config
 func (c *Config) ApplyDefaults() error {
 	if err := c.Prometheus.ApplyDefaults(); err != nil {
 		return err
+	}
+
+	if c.Logs != nil && c.Loki != nil {
+		return fmt.Errorf("at most one of loki and logs should be specified")
+	}
+
+	if c.Logs == nil && c.Loki != nil {
+		c.Logs = c.Loki
+		c.Loki = nil
+		c.UsedDeprecatedLoki = true
 	}
 
 	if err := c.Integrations.ApplyDefaults(&c.Prometheus); err != nil {
@@ -74,7 +97,7 @@ func (c *Config) ApplyDefaults() error {
 
 	// since the Tempo config might rely on an existing Loki config
 	// this check is made here to look for cross config issues before we attempt to load
-	if err := c.Tempo.Validate(&c.Loki); err != nil {
+	if err := c.Tempo.Validate(c.Logs); err != nil {
 		return err
 	}